Add support for custom servo assembly commands
This makes users' extensions to the servo assembly firmware
more convenient to use, as it no longer requires "hijacking"
other parts of the protocol.

An API endpoint, `r.servo_board.custom_command("foo 1 2 3")` is
added. This sends a message to robotd with `custom-command`
set to the user-provided string (in this case "foo 1 2 3").
robotd then forwards this to the servo assembly as a command
in the form "custom foo 1 2 3". The servo assembly dispatches
this to the "custom" handler function, which is a stub in
which the user can add their functionality. Result lines from
the user's code are passed back via robotd to the API, where
they are returned as a list of strings.